Program 2-8 Program 2: C Program to display Hello World! One print statement, two seperate lines.
Program 3:
C Program to do addition of two integers.

Program 4:
C program to subtract two integers.

Program 5:
C program to multiply two integers.

Program 6:
C program to divide two integers.

Program 7:
C program to find the Area of a Rectangle.
Length is 5 and Breadth is 10. Find the Area.

Program 8:
C program to find the Area of a Triangle.
Base is 5 and Height is 10. Find the Area.